Use a relative path in call to pdftohtml.
pdftohtml will use the file path as the basis for generating the
image links in the generated HTML, so we want them to be relative,
not absolute. Ruby 1.9 tempfile.path will generate an absolute
path when given the current directory as the tmpdir, whilst 2.0
and 1.8 will use a relative one.