Avoid loading associations to check existence
Calling InfoRequestEvent#is_X? loads the associated model from the
database. We don't need to do this to check what 'type' the event is.

This commit first checks whether there is an association_id – if there
is then we know it's that 'type'. If there's no association_id, call the
association to check whether one has been added but not yet persisted.

Only check association if the object is a new record. This avoids
checking the association on a persisted record when we know it isn't
that type of event due to the database record not existing.