Make message field required when redelivering incoming messages
The admin form for incoming messages allowed the user to redeliver
a message to a different request(s), but it didn't check that the user
actually supplied any request identifers before processing it. The
last step of the processing was to delete the existing message,
regardless of whether it had been sent anywhere new.

This commit changes that, so that the field is HTML5 required and the
controller checks before doing any more processing.