Use a string for spec descriptions.
"The semantics of `describe <a Symbol>` are changing in RSpec 3. In RSpec 2,
this would be treated as metadata, but as the first `describe` argument,
this will be treated as the described object (affecting the value of
`described_class`) in RSpec 3. If you want this to be treated as metadata,
pass a description as the first argument."