Replace pending with skip to preserve spec behaviour in rspec 3.
"Pending examples are now run to check if they are actually passing.
If a pending block fails, then it will be marked pending as before.
However, if it succeeds it will cause a failure. This helps ensure
that pending examples are valid, and also that they are promptly
dealt with when the behaviour they describe is implemented.

With this change, passing a block to pending within an example no
longer makes sense, so that behaviour has been removed."