Move spam message handling to InfoRequest
Requests can have instance-specific message handling, so allow this to
happen before we attempt to do any spam filtering. The instance handling
might reject the message anyway.

The ‘should purge the downstream cache when mail is received’ spec was
failing because the holding pen is now being called (and created, if
this is the first time the holding pen has been initialised) in
InfoRequest#receive. This seemed to create a PurgeRequest for the
holding pen instead of the request under test. The solution was to
ensure the holding pen already exists before this spec. We discussed in
standup that its probably best to remove PurgeRequest in the future.