Extract spam checking to class
Still need to keep the hack to prevent a stack level too deep error
where messages redirected to the holding pen are redirected back to the
holding pen continuously.

stub_const is used rather than stubbing the AlaveteliConfiguration
directly because the constant assignment seems to happen before the
stubs are initialized.