add "simple" example of censor rule (word match and regexp)
note this doesn't cover the caseinsentive example well:
as @schlos asked https://github.com/mysociety/alaveteli/issues/2056#issuecomment-93829106
about how to add the "options" that regexp users might expect.
I don't think the Nicargauan example really covers this
because it adds the regexp within the source code.

If we can show it in the admin interface, that would be
great but I'm not sure how to do that: as it happens,
I've simulated

* case-insenstitive (by using character class [Aa] on one letter)
* multiline (by using whitespace \s)

both of which would be better if I was bolder about saying
"this is how to do it."  But I'm not convinved it's the best
way... is it?