Implementation of @lizconlan's original fix for #2736.
Looking at this again, I think I was wrong in pushing back on this
original solution. The problem is that strings containing quotes etc,
interpolated into translated strings, get html escaped and marked as
`html_safe`. This is fine when they are immediately displayed in a
template as the `html_safe` marker means they don't automatically get
escaped again. However, when the resulting strings are stored in a `PostRedirect`
`reason_params` field as YAML and then pulled out again for display
on the signup page, the html_safe attribute is lost, so the string
is double escaped.

The `web` param of a track is only ever displayed via a `PostRedirect`
on the `sign` template, I think it's better to mark interpolated values
as safe originally, so that the whole string gets escaped in the template
like other non-translated strings. This won't fix any existing `PostRedirects`,
but these will be deleted in 3 months. It doesn't require us to make
special use of `html_entities`, so is more consistent with what's going
on in general.

Sorry for the bad review @lizconlan!