Fix dynamically created translation
Translation was constructed in two parts. Note that the @title ivar in
the view is still dyamically constructed, but it should be in a way that
makes sense.

For example, a translator in a RTL language could translate "Public
authorities - {{description}}" as "{{description}} - Public
authorities". The hyphen is siginifying a breadcrumb in the page title,
so the existing translation for {{description}} should still work.