Add rake task for checking required help templates and sections.
This has a list of templates that are expected to exist and
sections that are referred to in Alaveteli - either through links
or redirects. We could generate the list automatically, but it's
a bit tricky to accomodate things like linebreaks in the source code
so this seems like a reasonable first step.