Upgrade compass, sass and foundation-rails
Moving to foundation-rails 6.x.x.x introduces the next version
of the Foundation framework so that has been avoided here.

Moving to sass 3.4.22 brings up deprecation warnings for Ruby 1.9.3
which break specs and could potentially appear in generated content.
(Moving beyond 3.4.22 loses Ruby 1.9 support so not currently viable.)
Without the need to pin it, sass should not be listed separately in
the Gemfile as it is installed as a dependency of Compass, Bootstrap
and Foundation.