Redirect to frontpage when no 'comment' param supplied.
Given that the comment field is hidden visually, I think it's OK
to expect it to be supplied as blank - the real life cases where it
isn't supplied at all are spam. Also now handles the case where no
params are supplied with a POST request by rendering the form.