Add accepted formats to frontpage and blog
Bots make requests for weird formats, presumably to poke at security
holes. Here's an example:

    $ curl -I -H "Accept: application/vnd.ms-excel" \
        -H "Content-Type: application/vnd.ms-excel" \
        -X GET \
        http://10.10.10.30:3000?q=user

This was causing an ActionView::MissingTemplate error.

This change results in a 406 Not Acceptable response, saving us a bunch
of exception notification mail. We should probably do this elsewhere,
but / and /blog seem to be the most common targets.

See http://www.justinweiss.com/articles/respond-to-without-all-the-pain/
for a bit more detail.