Allow MailServerLog#line to be decorated
Wrap the line attribute in a decorator appropriate to the MTA, providing
extra functionaility when interacting with the line.

We could _always_ decorate the line, but to maintain compatibility we
might want to allow String methods to be forwarded to the line attribute
of the decorator. This is fine, but for now I didn't want to go in to
writing the extra test cases in case fleshing out PostfixLine brings up
any gotchas.