Consider users with comments or tracks as genuine
Much more likely to be genuine users, so just skip them for now.

Probably want to do something a little more intelligent, e.g. give these a
negative value in the score mappings.

squash! Consider users with comments or tracks as genuine

This is a bit of a temporary fix, because ideally it would be better to
filter users with comments _before_ spam checking them. This will be
easier when we add a counter cache to transaction associations:
https://github.com/mysociety/alaveteli/issues/3283