Cache outgoing message delivery status
A cache of 6 hours is probably the best between update frequency and
minimising page load. This will effectively minimise the calculation to
4 times a day.

A shorter cache (15 minutes, 1 hour) wouldn't catch the long tail of
requests that only get a handful of visits per day.

A longer cache would result in a long delay between changing to an
updated status.