Add UTM params to similar requests sidebar
Tracks clicks through to requests from the Similar Requests part of the
sidebar on RequestController#show. Does not track clicks to bodies or
users because the helper methods don’t allow optional parameters.

The `request_user_link` helper is particularly complex due to it picking
a User Name if the request is external.