Convert Fixnum to ActiveSupport::Duration for query
`InfoRequest.find_old_unclassified` called `#days` on `age_in_days` to
convert it to an `ActiveSupport::Duration` [1]. I think the gist is that
it converts `3` in to `259200` seconds ago, or something like that.

`OLD_AGE_IN_DAYS` already has `#days` applied to it, hence the longer
conditional, rather than calling `#days` in the query interpolation.
Doing so extends the old age:

    InfoRequest::OLD_AGE_IN_DAYS
    # => 1814400
    InfoRequest::OLD_AGE_IN_DAYS.days
    # => 156764160000

[1]
http://api.rubyonrails.org/v3.2.22/classes/Numeric.html#method-i-days