Use constant to show Alaveteli version instead of git tag
This fixes a bug when Alaveteli is deployed via Capistrano because tags
aren't pulled with Capistrano. It also makes the display of this more
consistent as this is what's used in version.json.

I've removed the link to GitHub because it no longer makes sense to do
that as you may not have pushed tags to your Alaveteli fork on GitHub.