Close table cell tags
When rendered, these unclosed table cells create an extra cell. You
don't see this visually unless the table has borders or something.

I noticed this problem on Right To Know where table styles leaked in
from elsewhere recently.

This commit closes the <td> tags.

Would this table be better as an ordered list anyway? Seems like a lot
of markup to achieve not that much ¯\_(ツ)_/¯.