Don't raise an exception when getting missing favicon in development
By default Alaveteli raises exceptions in development, even for 404 Not
Found errors, such as when the browser tries to get the favicon from the
default location.

Normally this isn't a problem but when you're using better_errors it
only keeps the most recent exception around. So if better_errors loads
but then your browser makes another request for the favicon (which
throws an exception) the REPL doesn't work.

This change fixes that by allowing requests for "/favicon.ico" in
development to display the usual error instead of throwing an exception.