Add ability to search for requests made to tagged public bodies
In aea54a4 we copied the suggested way this search could work from #24
but in implementing it I realised that having
"tag:some_tag variety:sent" makes it ambiguous what the tag is on - the
request or the body?

So in this commit I've decided to create a new term for public body tags
on info_request_events in the same way that the tags from the
info_request is exposed.

To search for request with a particular public body tag, you use
"request_public_body_tag". It's a bit verbose but I think clarity is far
more important than brevity here.

Also adds documentation about request_public_body_tag to advanced search page.

Fixes #24.