Inherit from main app ApplicationController and proxy path helpers
We need to inherit from ApplicationController in order to get the helper
methods like site_name, and we need to pass path helpers through to the same
class so that we can use the main application's layout views which use them
too.

See: http://candland.net/2012/04/17/rails-routes-used-in-an-isolated-engine/

This doesn't seem totally right, but it allows us to use the application
layout for now.