WIP: Add a professional page you have to log into
This is trying to reuse Alaveteli's Auth mechanisms for protecting an
Alaveteli Professional page. Currently it's not working very well because the
code it uses from Alaveteli (The authenticated? method) has to redirect the
user to the login url, and that url helper isn't available from the engine.

I'm not sure why it's necessary for it to be accessible, since it's Alaveteli
code calling it, not AP code, and a similar pattern seems to work ok for
ActiveAdmin and other engines with pluggable authentication methods.