Add an Alaveteli::Features module using Rollout
This adds a new module Alaveteli::Features which contains a single method:
active?(:feature_name, user_instance) that lets you check if a given feature
is activated (optionally for a given user). Available feature names are
exposed through the Alaveteli::Features::FEATURES constant and as individual
constants on the module. The module is defined and configured in
config/initializers/features.rb

Underneath this uses https://github.com/fetlife/rollout to manage
activating/deactivating features, but with an adapter class so that we can
store feature config in the database rather than in Redis. This requires a
new model to store these configs in.