Match status text color to icon’s color
In a1e9018e315d751cabd74df7a687ccedad8f27a7 the status icons were
updated, but the color of the text that goes with it wasn't.
This leaves a messy effect that reduces the clarity of the status.

This commit updates the colors to match the icons, based on the text
color used in the WhatDoTheyKnow theme (the original source of the
icons):
https://github.com/mysociety/whatdotheyknow-theme/blob/master/assets/stylesheets/responsive/custom.scss#L39