Amend scope for Comment.visible to exclude InfoRequest.is_public select
The select on InfoRequest.is_public was added in 9e022ada so that
the InfoRequest records returned would not be read-only (which they
otherwise would be due to the use of :join rather than :include to
get the outer join for non-embargoed requests). If carried over here,
it results in the InfoRequest, not the Comment, being returned.