Skip public body censor rules when there is no body
Previously this only skipped when the request was a batch request template, because
that is the only legitimate time that a request can have no body. However, this
didn't think about the fact that this method is called during validation of
outgoing messages, so if you create a request without a body AND an outgoin
message for it, you'll get an error rather than a validation error.