Use local time wherever possible.
Model times will be converted to the time zone which is set on the
application automatically by ActiveRecord (and saved in UTC in the
database). Times with zones (e.g. as returned by Time.zone.now) will
be automatically converted to UTC in queries. Time.now returns
the system time with no zone, so will not be converted in queries.

As demonstrated by the spec in this commit, comparisons in code
between Time.now and times calculated using model fields can produce errors.
It's safer to use Time.zone.now consistently.

More info on how Rails handles timezones at:

https://www.varvet.com/blog/working-with-time-zones-in-ruby-on-rails/