Clarify admin info message
9e215bd added a note to address the problem where we can't fall back to
the default locale's request_email attribute. What I didn't understand
at the time was that it _does_ fall back when there are _no_
translations.

This is because when there are no translations for a locale, there's no
PublicBody::Translation record. When there are translations, it just
finds an empty field from the record and returns that.

Ticket https://github.com/mysociety/alaveteli/issues/3134.