Pass public body when calling create_request, update xapian index
Should solve transient CI failures where public body not found
when creating request. The cause of this is the body not being in
the xapian index (depending on what other specs have previously
run). So, we're now passing the public body to the create_request
method, and updating the xapian index to ensure it's there.

Change body name, as since 9f1fe3ca, searching for a name with
fewer than 3 characters in the last word causes zero results to
be returned.