Fix ZIP download of unclassified requests
The "state transition variables" were not set, causing the describe
state form to get rendered without any state transitions.

When downloading a Zip, `@transitions` is nil, causing an `undefined
method `[]' for nil:NilClass`, so I've added a `nil` check in
`state_transitions_empty?`.

Then set instance variables in `assign_variables_for_show_template` so
that we don't render the describe state form at all when rendering to file.

I've tidied up the setting of some of these as it was pretty
complicated, and wrapped them all in `!!` so that you're guaranteed a
Boolean. While not strictly necessary, it makes it easier to test
because some conditions caused the expected value to be `nil` (e.g. `nil
&& false # => nil`).