Remove intersection of div and form elements.
The integration spec for changing an embargo was failing, and I
think the reason was that the houdini elements introduced to hide
this form when not in use resulted in a div - "houdini-target" -
that started inside the form and ended outside it. I don't think
that's valid CSS. By moving the div to encompass the form, the
issue seems to have been resolved (although it does leave the other
houdini elements outside a form context).