[DROP] Explicitly set HABTM join table
We don't need this.

Took a few reads through the explanation for it to click. What its
saying is that (for the example provided) the default used to be
`catalog_categories_catalog_products`, but will now be
`catalog_categories_products`.

[WIP] Explicitly set HABTM join table

Not sure if we need this but can’t run the tests yet…

> Rails 4.0 has changed to default join table for
> has_and_belongs_to_many relations to strip the common prefix off the
> second table name. Any existing has_and_belongs_to_many relationship
> between models with a common prefix must be specified with the
> join_table option. For example:
>
>   CatalogCategory < ActiveRecord::Base
>     has_and_belongs_to_many :catalog_products,
>       join_table: 'catalog_categories_catalog_products'
>   end
>
>   CatalogProduct < ActiveRecord::Base
>     has_and_belongs_to_many :catalog_categories,
>       join_table: 'catalog_categories_catalog_products'
>   end
>
> – http://edgeguides.rubyonrails.org/upgrading_ruby_on_rails.html#
>     upgrading-from-rails-3-2-to-rails-4-0

squash! [WIP] Explicitly set HABTM join table

We don't need this.

Took a few reads through the explanation for it to click. What its
saying is that (for the example provided) the default used to be
`catalog_categories_catalog_products`, but will now be
`catalog_categories_products`.