Submit pro batch authority search via ajax
This handles submission of the form and clicks on the pagination links
making both of those go via AJAX instead of reloading the page. Also
applies a really basic class to the whole container and to pagination
links whilst an AJAX request is happening, so that later we can style
it appropriately to show the user what's going. In the interim,
buttons to add authorities and pagination links are disabled during
loading.