Remove test_unit gem
This was causing an rspec example that did:
expect(response).not_to render_template(...)
to fail when it shouldn't have been. As per:
https://github.com/rspec/rspec-rails/issues/1539
this is because of test_unit also having a render_template
matcher and this interfering with the negation. I can't
see any reason why this gem is needed now, as we run all our
specs through RSpec.