Add javascript for submitting the 'Remove body' buttons
This required some reworking of the load order of javascript files, because
both the add and remove button components relied on their parent modules, so
those both had to be initialised first. I could have solved this with
something like RequireJS but I wanted to keep this as simple as possible for
now.