Fix incorrect counter cache after updating association
Rails 4 introduces a bug that causes counter_cache code to be
fired twice when updating an association [1].

Its still an open issue, so instead of hacking around with Rails
internals, manually reset the counter cache when changing the associated
record. Also hacks in a forced update of the associated public body caches.

Note that some specs expect change from `nil` to `1` because the counter cache
column doesn't have a default value set.

[1] https://github.com/rails/rails/issues/10865