Set confirmed_not_spam to true in User factory
Most of the time we want to test against genuine users. It can be
confusing when `FactoryGirl.create(:user)` ends up failing in specs
because of spam prevention measures.

Though in reality a user's `confirmed_not_spam` defaults to `false`, its
much easier to assume factory users are genuine. If we're testing anti-spam
features we can build an unconfirmed user factory easily enough.