Make sure batch search results are reloaded after first add
This is less than ideal, but because the initial search results render
'create' forms for the draft request, they can't be used later to add
more bodies onto a request. Swapping around all the necessary html in
javascript would be difficult, so instead we resubmit the search form
once a body is added for the first time. After that we can keep using
the initial HTML because everything is set up.