Get a clean copy of xapian index before any specs run, leave it in the same state after specs.
We use `get_fixtures_xapian_index` to know before the specs run what's in the xapian index.
The first time `get_fixtures_xapian_index` is called in a spec run, it creates a
xapian index with whatever data is in the database at the time. For subsequent calls,
it takes a copy of the index generated on the first call. We want the result of
`get_fixtures_xapian_index` to always be a clean index copy with only spec data in it,
so we need to call it before the `let!` calls. Otherwise this copy will have the authorities
created in the `let!` calls in it. We also need to make sure that the authorities created
in it don't accumulate between specs. If that happens, ActsAsXapian ignores them when
populating models for a result set, as the models themselves no longer exist in the
relational database, but they gradually push the real results out of the 25 returned. Also,
the next bit of code to call `get_fixtures_xapian_index` would be getting an index with
unexpected models in it.