Test whether feature flipper tables exist better
At the point when we were trying to test them, ActiveRecord::Base.connected?
would always return false, because you have to actually try to establish a
connection and use it first. Instead, we now try to establish a connection
and then check for the tables, no need to test for connected? at all. A
rescue block allows us to deal with any issues regarding a missing db etc
at the same time.