Further fix for MailServerLog#delivery_status
Because we're calling #save after rescuing from an ArgumentError,
delivery_status was returning true instead of the delivery status.

Use the serialiser after this to ensure the DeliveryStatus is loaded.

This was missed because even `#update_attribute` does weird things with the
serialiser, so set the value manually with SQL.

Also tidy up the happy-path so that set_delivery_status is only called
if the delivery_status attribute is not present for a bit clearer code.