Make sure batch show action lists embargoed requests for the right people
Pros should be able to see embargoed requests as part of a batch.
Note: this is currently assuming that all requests in a batch are the
same, i.e. either embargoed or not, so it will break if we start allowing
people to selectively embargo or publish things. We'll need to make some
changes to avoid the latter being possible.