Use helper to display distance of time
Using `#inspect` was causing the following error on IPV:

    An I18n::InvalidLocale occurred in admin_general#index:

    :en is not a valid locale

    app/views/admin_general/index.html.erb:203:in `_app_views_admin_general_index_html_erb___4193968884139696355_74739300'
    app/controllers/application_controller.rb:113:in `record_memory'
    lib/strip_empty_sessions.rb:13:in `call'

Wrapped in a `with_locale` block so that it uses English if possible;
otherwise it reverts to the default locale. I think the error above must
be caused by Rails trying to default to `:en` when it isn't a configured
locale.