Enforce available locales
This option was added as a security measure, to ensure user input cannot be
used as locale information unless it is previously known. Therefore, it's
recommended not to disable this option unless you have a strong reason for
doing so.

http://edgeguides.rubyonrails.org/upgrading_ruby_on_rails.html#i18n-enforcing-available-locales

This was originally introduced in
https://github.com/mysociety/alaveteli/commit/5b709d61c1c897f10d9900018fecca4c10731673
to match previous behaviour and to avoid a deprecation warning. I can't
see any deprecation warnings when setting it to `true`.

Further reading here:

* http://stackoverflow.com/questions/20361428/rails-i18n-validation-deprecation-warning
* https://github.com/svenfuchs/i18n/commit/3b6e56e06fd70f6e4507996b017238505e66608c