Parse correct MTA ID from syslog-format logs
Fixes https://github.com/mysociety/whatdotheyknow-theme/issues/392.

Syslog prefixes the exim log line with date, hostname and process info,
so shifted the exim message id from the 4th element to the 9th after
splitting on whitespace.

The message id is a consistent format [1] so we can parse it out
reliably with a regex.

* The first `#compact` makes sure that we're not trying to match against
  `nil` values.
* The second `#compact` makes sure that no `nil` values are returned if
  a message id was not found.

[1]
http://www.exim.org/exim-html-current/doc/html/spec_html/ch-how_exim_receives_and_delivers_mail.html#SECTmessiden