Update exception_notification gem
Updated to get a version which supports the new ActionMailer::MessageDelivery
methods. Pinned to < 4.1.2 as that's when Ruby 1.9 support is dropped

Here be dragons - previous commit message (6f4931de) reads:

"4.1.0 has a bug in it. The bug gets fixed in a later version that
doesn’t have Ruby 1.9.3 support."

Although the last update commit for this gem (cb9b2a0) suggests there
have been no further updates for Ruby 1.9.3, it looks as though the
critical bug (triggered by a change to how monkeypatching is handled
within Mailer) from 4.1.0 is fixed in 4.1.1 which is the actual last
version with Ruby 1.9.3 support.

Local testing confirms that version 4.1.1 successfully creates an
exception notification mail whereas 4.1.0 fails immediately per:
https://github.com/smartinez87/exception_notification/issues/270