Make AlaveteliPro::AccountMailer#account_request return a Mailer object
...and move the responsibility of sending the mail to calling method
(in this case AlaveteliPro::AccountRequestController#create)

This is for Rails 4.2 compatibility as in enabling delayed mail sending,
Mailer instance methods calls are now deferred until the `deliver`
method has been called. So if we nest our call to `deliver` inside
an instance method it can never be reached once we've upgraded to 4.2.

http://edgeguides.rubyonrails.org/upgrading_ruby_on_rails.html#action-mailer