Update bad incoming email processing specs after switch to user.notify
I've kept this as a separate commit to make it easier to review and clearer in
intent. Various specs in the incoming mail handling and info request spam
detection were relying on the fact that receiving a new response sent an email
(and in parts testing that behaviour) which is no longer the case.

I don't think they need to test that, it is a separate piece of code and is
tested separately, so rather than testing the equivalent new code, which is
that a Notification object is created, I've just removed those lines from the
specs.