Improve PublicBody#url_name validation specs
* Add presence validation specs
* Move and improve style of uniqueness validation spec
* Move specs for creating slugified `url_name`

The slugifying of `url_name` doesn’t really belong in the `#url_name`
spec block, as it’s being changed in `#name=` and `#short_name=`. This
could do with a bit of tidying up in itself, so just moving it to here
for now so that the behaviour is documented in a reasonably obvious
place.