Improve PublicBody#first_letter specs
Describes the main behaviour.

We _should_ add validations to this, but because the value is getting
set through a `before_save` callback it means that we’d have to change
to a `before_validation` callback so that it doesn’t cause records to
be invalid at validation.

I don’t see why we don’t set `#first_letter` in `#name=`.