Update the RequestSummary every time we log an event
Not every event logging will save the request, but it could update the
calculated status, so we need to update the request summary in order to
reflect that. This fixes a specific issue we had where requests being marked
as overdue by a cron job weren't triggering the necessary update to the
RequestSummary.

See: https://github.com/mysociety/alaveteli-professional/issues/306