Use partial to render flash message
Assigning HTML to the partial does not work as the `html_safe` marker
is not persisted across a redirect.

Under the hood, `html_safe` converts the `String` to an
`ActiveSupport::SafeBuffer`. The flash gets serialized to JSON, so the value
gets deserialized to a `String` rather than the `SafeBuffer`. The value then
gets escaped.