Validate that request summaries are unique per request
As part of this, I had some troubles with the request summary factories
(again) because they were creating invalid summaries. I've moved the
slightly tricky stuff that they do to get a clean slate of one request
with one summary into an after(:build) rather than a before(:create)
because that means it works for any type of factory usage, not just
FactoryGirl.create().