Use request that's passed in when updating summaries, not local one
RequestSummary#update_from_request is called by code that has access to
a guaranteed-most-up-to-date copy of the request, but it discards it
and accessess it's own copy through .summarisable instead. In the case
of status updates through InfoRequest#set_described_state (and perhaps
in other cases) this meant that the summary wasn't seeing changes to
its request's status, because it's local copy was cached. I could have
forced a reload in the method but it seemed better to utilise the copy
of the request that was already present by passing it in, making the
update and create methods more consistent in the process.

For mysociety/alaveteli-professional#310