Use the translated PublicBody name for sorting on associations
In preparation for dropping translated columns.

This syntax ensures that only the current locale is used in the SQL
join to avoid loading translations for all locales and, as previously,
uses reorder to ensure this is the only sort order being used.

Once we upgrade to Globalize 5.0.1 (requires Rails 4.2) we could update
this to:

has_and_belongs_to_many :public_bodies, -> {
  with_translations([I18n.locale]).
    reorder('public_body_translations.name asc')
}

as it looks as though the current problem this causes - an attempt is
made to assign locale to the join table when assigning public_bodies,
raising an error - has been resolved from 5.0.1 onwards.

Fixes transient spec failures related to admin_raw_email_controller
and public_body.rb which appear to be to do with using `includes`
in a HABTM relationship - may introduce eager loading (possibly
dependent on spec run order)

Example failures:
https://travis-ci.org/mysociety/alaveteli/builds/247451498