Add a temporary sleep in batch request sending
Because requests were sent in a separate loop to their creation, we had to
refactor the sending code a little bit. Sleeping before sending would have
introduced the potential for requests to be indexed before they were sent,
which would cause an error due to the lack of an initial event. Instead, each
request is created and the batch updated in a transaction, then it's sent,
after which we can sleep. We update the batch this way so that if any request
errors, the batch is already marked as sent, which means it won't be
automatically resent. Before, this happened at the end of a transaction where
all the requests were created, but that didn't give us a place to insert a
sleep easily.