Remove unconventional constraints from public_bodies
Some columns set `NOT NULL` with a default of `""`. Our best guess is
that this was done for consistency, as submitting an empty form field
will result in the column being set with `""`. We now use
`strip_attributes`, which converts `""` to `nil` before persisting the
value, so this seems redundant.

In this migration we replace empty values with `nil` to make existing
records consistent with records created from this point onwards.

Need to exclude `request_email` from the initial strip otherwise it gets
stripped, making the next line that targets it redundant.

Don't expect PublicBody#short_name to set a default value