Remove fragment caching of delivery status
In some cases the delivery status can get cached, but on clicking the
dropdown the info is refreshed and shows a message not inline with the
cached status.

The front end caching can now be removed because the calculation is cheaper
since https://github.com/mysociety/alaveteli/pull/3476.

Fixes https://github.com/mysociety/whatdotheyknow-theme/issues/404.