Ignore line-length for rspec it/describe/context
Splitting the string between an `it` and `do` breaks the ability to run
specs by the correct line number (e.g. `/path/to/spec.rb:7`).

The printed line number is that of the `it`, but the line number
necessary to run that spec is the line with `do` – usually a line below.
This can make life confusing when you run what rspec reports as the
failing test and it immediately passes.