Tidy the import step
I found the original really hard to follow, and missed that it was
yielding a block. I thought the block was iterating the
`default_available_locales`.

This builds the importer arguments outside of the method call so that
the block is more obvious. It also sets the assigned variables on the
preceeding line so that its clearer what we're getting back.