Menu side set in css by page context, not in markup
The action menus at the top and bottom of the page were
either both projecting to the left (on embargoed requests)
or to the right (on public requests). This commit alters
the css so that the left projection is applied only to the
action menu in the header (which is positioned to the
right hand side of the request).

Given that this is possible, I don't think there's a need
to specify in the element classes whether they are right
or left positioned.