Only update summary in callbacks, always reload the request
When `create_or_update_request_summary` is called in the context of
a callback, we can't be sure that we don't have a stale copy of the
request. This can happen specifically when the update to the request
was done in a complex transaction that referenced some other instance
of the request before the one that gets the changes we care about.

We don't want to reload on every call to `create_or_update_request_summary`
as it can also be called from `log_event`, where changes to the request
are not yet persisted, so a reload would lose them.