Prevent unintentional attribute updates in getter methods
Until Rails 4.2, creating a temporary variable with the same name as the
underlying attribute inside the getter method of an ActiveRecord model
was treated as a new variable (as intended here). Instead it is now
acting as a reference to the named attribute causing unnecessary lookups
and, more importantly, writes to take place. (Note: the write may not
be apparently until save is called.)

In this case, we were mistakenly updating the stored value of the user
name which caused the url_name to be updated as well, which was caught
by the spec "banning the user ... does not change the URL name":
./spec/models/user_spec.rb:59

* Updates code to use non-clashing variable names for:
  * User#url_name
  * InfoRequest#title
  * MailServerLog#line
* Adds a new spec to check that User#name is not changed by banning