Use default local gem dir, add to path in daemons and crons
This is necessary because we now install bundler using rubygems
for older OSes. When installing to our own GEM_HOME, cron jobs
and daemon files weren't getting GEM_HOME and so weren't able
to load bundler correctly. If we use the standard local gem
dir for our installation user, it's already in rubygems'
default path list, and we only need to give crons and
daemons the right PATH prefix.