Make AlaveteliLocalization return locales as strings
It felt like keeping locales as symbols would be more efficient but it
turns out that for most of the places we use them - in URLs, in form
field names, in SQL queries... we need to coerce them to strings, losing
the tiny efficiency saving of symbols and making the code look untidy.
Strings it is then.

The one notable exception is going to be translation.locale (from
Globalize) which returns a symbol:
https://github.com/globalize/globalize/blob/v4.0.3/lib/globalize/active_record/translation.rb#L27-L30